i82: duplicated JavaHelpersSpec.scala removed
